Viser en info-side fremfor å redirecte ved requests fra editoren
