+++ src/components/PageBase.tsx
@@ -114,12 +114,14 @@ export const fetchPageProps = async (
         return errorHandler(content);
     }
 
-    const redirectTarget = getTargetIfRedirect(content);
-    if (redirectTarget) {
-        return redirectProps(
-            getRelativePathIfInternal(redirectTarget, isDraft),
-            content.data?.tempRedirect
-        );
+    if (!isDraft) {
+        const redirectTarget = getTargetIfRedirect(content);
+        if (redirectTarget) {
+            return redirectProps(
+                getRelativePathIfInternal(redirectTarget, isDraft),
+                content.data?.tempRedirect
+            );
+        }
     }
 
     return {
