+++ src/components/pages/redirect-page/RedirectPage.less
@@ -1,0 +1,11 @@
+@import (reference) '../../../global';
+@import (reference) '~nav-frontend-typografi-style/src/index';
+
+.redirect-page {
+    @paddingVertical: 2rem;
+    @paddingHorizontal: @paddingVertical * 1.62;
+    width: 100%;
+    background-color: white;
+    padding: @paddingVertical @paddingHorizontal;
+    border: 4px solid @navGronnLighten60;
+}
